comparison fix

+++ src/engine/server.cpp
@@ -988,7 +988,7 @@ int updatetimer(bool limit)
         int cursecs = (totalmillis-lastsec)/1000;
         totalsecs += cursecs;
         lastsec += cursecs*1000;
-        ifserver(maxruntime && !shutdownwait && totalsecs >= maxruntime)
+        ifserver(maxruntime && !shutdownwait && totalsecs >= uint(maxruntime))
         {
             server::srvoutf(-3, "\fymax run time reached (\fs\fc%s\fS), waiting for server to empty", timestr(maxruntime*1000, 4));
             shutdownwait = totalmillis;
